Added various objects
