+++ index.html
@@ -11,7 +11,6 @@
   </head>
   <body>
     <canvas id="demo2" width="800" height="600"></canvas>
-    <!-- <canvas id="demo2" width="800" height="600"></canvas>
-     NOTE: SCRIPTS WILL BE AUTOMATICALLY ADDED BY WEBPACK BELOW THIS COMMENT -->
+    <!-- NOTE: SCRIPTS WILL BE AUTOMATICALLY ADDED BY WEBPACK BELOW THIS COMMENT -->
   </body>
 </html>
