+++ src/engine/plugins/three/Factory.ts
@@ -1,0 +1,10 @@
+import { ISim3D } from "../../interface/ISim3D";
+import { SimulatorConfig } from "../../interface/SimulatorConfig";
+import { Sim3D } from "./Sim3D";
+
+export function createThreeSim3D(
+  canvas: HTMLCanvasElement,
+  config?: SimulatorConfig
+): ISim3D {
+  return new Sim3D(canvas, config);
+}
